refactor: move the api classes over to the new module
Move the api classes to the new datasource module
