+++ app/build.gradle
@@ -33,6 +33,10 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
+
+    implementation project(":feature:pusher")
+    implementation project(":library:datasource:pwpush")
+
     implementation "androidx.activity:activity-ktx:1.2.3"
     implementation "androidx.appcompat:appcompat:1.3.0"
     implementation "androidx.constraintlayout:constraintlayout:2.0.4"
@@ -43,8 +47,6 @@ dependencies {
     implementation "com.chesire.lintrules:lint-xml:1.2.6"
     implementation "com.google.android.material:material:1.3.0"
     implementation "com.squareup.okhttp3:okhttp:4.9.1"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.1"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.1"
 
     testImplementation "junit:junit:4.13.2"
