+++ app/src/main/java/com/chesire/pushie/MainActivity.kt
@@ -9,8 +9,8 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.chesire.pushie.api.PasswordPusher
 import com.chesire.pushie.databinding.ActivityMainBinding
+import com.chesire.pushie.datasource.pwpush.remote.PasswordPusher
 import com.chesire.pushie.extension.closeKeyboard
 import com.google.android.material.snackbar.Snackbar
 import okhttp3.OkHttpClient
