+++ library/datasource/pwpush/build.gradle
@@ -35,6 +35,9 @@ dependencies {
     implementation "androidx.core:core-ktx:1.5.0"
     implementation "com.chesire.lintrules:lint-gradle:1.2.6"
     implementation "com.chesire.lintrules:lint-xml:1.2.6"
+    implementation "com.squareup.okhttp3:okhttp:4.9.1"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.1"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.1"
 
     testImplementation "junit:junit:4.13.2"
 }
