+++ library/datasource/pwpush/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.chesire.library.datasource.pwpush" />
+<manifest package="com.chesire.pushie.datasource.pwpush" />
