+++ library/datasource/pwpush/src/main/java/com/chesire/pushie/datasource/pwpush/PWPushRepository.kt
@@ -1,0 +1,5 @@
+package com.chesire.pushie.datasource.pwpush
+
+class PWPushRepository {
+
+}
